+++ test/test_source/hover/functions.f90
@@ -1,0 +1,10 @@
+! simple function
+function fun1(arg)
+    integer, intent(in) :: arg
+    integer :: fun1
+end function fun1
+
+! function with type on definition, implied result
+integer function fun2(arg)
+    integer, intent(in) :: arg
+end function fun2